@@ -6,7 +6,7 @@
   "license": "LicenseRef-LICENSE",
   "version": "0.0.1",
   "engines": {
-    "node": ">=4.0.0",
+    "node": ">=5.0.0",
     "npm": ">=3.0.0"
   },
   "repository": {
@@ -21,8 +21,7 @@
     "alt": "~0.17.8",
     "bigi": "1.1.0",
     "bs58": "2.0.0",
-    "bytebuffer": "^5.0.0",
-    "immutable": "^3.7.5",
+    "immutable": "^3.8.1",
     "lodash": "^3.10.1",
     "lzma": "2.1.6",
     "numeral": "^1.5.3",
@@ -35,18 +34,16 @@
   "devDependencies": {
     "babel": "^5.8.29",
     "babel-core": "^5.8.33",
-    "babel-loader": "^5.3.3",
     "babel-runtime": "^5.6.15",
-    "babel-plugin-react-transform": "^1.1.1",
     "coffee-loader": "~0.7.2",
     "coffee-script": "~1.8.0",
     "coffeeify": "~0.7.0",
     "fake-indexeddb": "^1.0.3",
-    "mocha": "~2.2.5",
+    "mocha": "^2.5.3",
     "mocha-traceur": "~2.1.0",
     "node-fetch": "^1.3.1",
     "node-libs-browser": "~0.5.2",
     "node-localstorage": "~0.3.4",
-    "traceur": "0.0.92"
+    "traceur": "0.0.110"
   }
 }
